+++ visqol-rs/src/speech_similarity_to_quality_mapper.rs
@@ -17,14 +17,14 @@ impl SpeechSimilarityToQualityMapper {
 }
 
 impl SimilarityToQualityMapper for SpeechSimilarityToQualityMapper {
-    fn predict_quality(&self, similarity_vector: &[f64]) -> f32 {
-        const FIT_PARAMETER_A: f32 = 1.155_945_5;
-        const FIT_PARAMETER_B: f32 = 4.685_115_3;
-        const FIT_PARAMETER_X0: f32 = 0.765_523_2;
-        const FIT_SCALE: f32 = 1.2031409;
+    fn predict_quality(&self, similarity_vector: &[f64]) -> f64 {
+        const FIT_PARAMETER_A: f64 = 1.155_945_5;
+        const FIT_PARAMETER_B: f64 = 4.685_115_3;
+        const FIT_PARAMETER_X0: f64 = 0.765_523_2;
+        const FIT_SCALE: f64 = 1.2031409;
 
         let nsim_mean =
-            (similarity_vector.iter().sum::<f64>() as f32) / (similarity_vector.len() as f32);
+            similarity_vector.iter().sum::<f64>() / (similarity_vector.len() as f64);
         let mos = math_utils::exponential_from_fit(
             nsim_mean,
             FIT_PARAMETER_A,
